Remove next image from theme
